Fixed startup race in transport adapters #2867
