+++ akka-remote/src/main/scala/akka/remote/transport/AbstractTransportAdapter.scala
@@ -2,16 +2,17 @@ package akka.remote.transport
 
 import scala.language.postfixOps
 import akka.actor._
-import akka.pattern.ask
+import akka.pattern.{ ask, pipe }
+import akka.remote.Remoting.RegisterTransportActor
+import akka.remote.transport.ActorTransportAdapter.ListenUnderlying
+import akka.remote.transport.ActorTransportAdapter.ListenerRegistered
 import akka.remote.transport.Transport._
-import akka.remote.{ RARP, RemotingSettings, RemoteActorRefProvider }
+import akka.remote.{ RARP, RemotingSettings }
+import akka.util.Timeout
 import scala.collection.immutable
+import scala.concurrent.duration._
 import scala.concurrent.{ ExecutionContext, Promise, Future }
 import scala.util.Success
-import scala.util.Failure
-import akka.remote.Remoting.RegisterTransportActor
-import akka.util.Timeout
-import scala.concurrent.duration._
 
 trait TransportAdapterProvider extends ((Transport, ExtendedActorSystem) ⇒ Transport)
 
@@ -121,7 +122,9 @@ object ActorTransportAdapter {
 
 abstract class ActorTransportAdapter(wrappedTransport: Transport, system: ActorSystem)
   extends AbstractTransportAdapter(wrappedTransport)(system.dispatcher) {
+
   import ActorTransportAdapter._
+
   private implicit val timeout = new Timeout(3 seconds)
 
   protected def managerName: String
@@ -150,3 +153,38 @@ abstract class ActorTransportAdapter(wrappedTransport: Transport, system: ActorS
   override def shutdown(): Unit = manager ! PoisonPill
 }
 
+abstract class ActorTransportAdapterManager extends Actor {
+  private var delayedEvents = immutable.Queue.empty[Any]
+
+  protected var associationListener: AssociationEventListener = _
+  protected var localAddress: Address = _
+  private var uniqueId = 0L
+
+  protected def nextId(): Long = {
+    uniqueId += 1
+    uniqueId
+  }
+
+  import context.dispatcher
+
+  def receive: Receive = {
+    case ListenUnderlying(listenAddress, upstreamListenerFuture) ⇒
+      localAddress = listenAddress
+      upstreamListenerFuture.future.map { ListenerRegistered(_) } pipeTo self
+
+    case ListenerRegistered(listener) ⇒
+      associationListener = listener
+      delayedEvents foreach { self.tell(_, Actor.noSender) }
+      delayedEvents = immutable.Queue.empty[Any]
+      context.become(ready)
+
+    /* Simple imitation of Stash. It is more lightweight as it does not need any specific dispatchers or additional
+     * queue. The difference is that these messages will not survive a restart -- which is not needed here.
+     * These messages will be processed in the ready state.
+     */
+    case otherEvent ⇒ delayedEvents = delayedEvents enqueue otherEvent
+
+  }
+
+  protected def ready: Receive
+}
